% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SuppFuns.R
\name{findroot_twisting_psitheta_equation}
\alias{findroot_twisting_psitheta_equation}
\title{find root of equation $phi'(theta) = x$ for Two-steps Importance Sampling}
\usage{
findroot_twisting_psitheta_equation(x, ck, pkZ)
}
\arguments{
\item{x}{threshold of portfolio, postive real value}

\item{ck}{expoure of obligors, vector in m dimenstion}

\item{pkZ}{default probability conditional on Z, postive real value}
}
\value{
root of $phi'(theta) = x$
}
\description{
find root of equation $phi'(theta) = x$ for Two-steps Importance Sampling
}

