% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NumExas.R
\name{GenerateLossProb}
\alias{GenerateLossProb}
\title{Generate Large Loss probability for Numerical experiments in OptCredRisk}
\usage{
GenerateLossProb(PlainMC_Sampling_Num, IS_Sampling_Num, d, m, pk, ck, A, b, x,
  y, DriftZ = "yes")
}
\arguments{
\item{PlainMC_Sampling_Num}{scenarios for MC simiulation, positive integer}

\item{IS_Sampling_Num}{scenarios for IS simiulation, positive integer}

\item{d}{number of risk factors, postive integer}

\item{m}{number of obligors, positive integer}

\item{pk}{default probability of obligors, vector in m dimension}

\item{ck}{expoure of obligors, vector in m dimenstion}

\item{A}{risk loading matrix, matrix in (m, d) dimension}

\item{b}{risk loading vector, vector in m dimension}

\item{x}{threshold of portfolio, postive real value}

\item{y}{set of threshold of portfolio, postive vector}

\item{DriftZ}{flag indicator on whether we need shift mean of risk factors}
}
\value{
Large Loss probability for both IS and MC
}
\description{
Generate Large Loss probability for Numerical experiments in OptCredRisk
}

