% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IS.R
\name{IS_TwoSteps_OneLoss_for_x}
\alias{IS_TwoSteps_OneLoss_for_x}
\title{single loss probability $P(L>x)$ by IS simulation}
\usage{
IS_TwoSteps_OneLoss_for_x(IS_Sampling_Num, m, d, pk, ck, A, b, x, y,
  DriftZ = "yes")
}
\arguments{
\item{IS_Sampling_Num}{scenarios for IS simiulation, positive integer}

\item{m}{number of obligors, positive integer}

\item{d}{number of risk factors, postive integer}

\item{pk}{default probability of obligors, vector in m dimension}

\item{ck}{expoure of obligors, vector in m dimenstion}

\item{A}{risk loading matrix, matrix in (m, d) dimension}

\item{b}{risk loading vector, vector in m dimension}

\item{x}{threshold of portfolio, postive real value}

\item{y}{set of threshold of portfolio, postive vector}

\item{DriftZ}{flag indicator on whether we need shift mean of risk factors}
}
\value{
average $L$, $P(L>X)$ and 95 percents confidence parameter $sigma$
}
\description{
single loss probability $P(L>x)$ by IS simulation
}

