% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SubSolver.R
\name{Loading_coeff}
\alias{Loading_coeff}
\title{find optimal risk loading coefficients}
\usage{
Loading_coeff(Zd, epsm, n, method = 0)
}
\arguments{
\item{Zd}{sampling loading risk vectors, matrix in (n,d) dimensional space}

\item{epsm}{sampling loading, vector in n dimension}

\item{method}{= 0, using projection method, method = 1, using SDP method}
}
\value{
list of loading matrix A and loading vector b
}
\description{
find optimal risk loading coefficients
}

