% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SubSolver.R
\name{Local_data}
\alias{Local_data}
\title{reformulate sampling data}
\usage{
Local_data(matZ, veceps, rownum, n)
}
\arguments{
\item{matZ}{randomly generated systematic risk factors, matrix in (n,d) dimensional space}

\item{veceps}{randomly generated idiosyncratic risk factors, vector in m dimension}

\item{rownum}{index of obiligor, postive integer}

\item{n}{number of scenarios, postive integer}
}
\value{
reformulated matrix (Z, eps_k) and x_k.e
}
\description{
reformulate sampling data
}

