% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SuppFuns.R
\name{Max_Jx_TwoSteps}
\alias{Max_Jx_TwoSteps}
\title{Maximum of approximation upper bound function J(x) for Two-steps Importance Sampling}
\usage{
Max_Jx_TwoSteps(d, m, ck, pk, A, b, x)
}
\arguments{
\item{d}{number of risk factors, postive integer}

\item{m}{number of obligors, positive integer}

\item{ck}{expoure of obligors, vector in m dimenstion}

\item{pk}{default probability of obligors, vector in m dimension}

\item{A}{risk loading matrix, matrix in (m, d) dimension}

\item{b}{risk loading vector, vector in m dimension}

\item{x}{threshold of portfolio, postive real value}
}
\value{
maximum of J(x)
}
\description{
Maximum of approximation upper bound function J(x) for Two-steps Importance Sampling
}

