% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MC.R
\name{PlainMC_LossSet}
\alias{PlainMC_LossSet}
\title{loss probability distribution $P(L>x)$ by MC simulation}
\usage{
PlainMC_LossSet(PlainMC_Sampling_Num, m, d, pk, ck, A, b, y)
}
\arguments{
\item{PlainMC_Sampling_Num}{scenarios for MC simiulation, positive integer}

\item{m}{number of obligors, positive integer}

\item{d}{number of risk factors, postive integer}

\item{pk}{default probability of obligors, vector in m dimension}

\item{ck}{expoure of obligors, vector in m dimenstion}

\item{A}{risk loading matrix, matrix in (m, d) dimension}

\item{b}{risk loading vector, vector in m dimension}

\item{y}{set of threshold of portfolio, postive vector}
}
\value{
list of average $L$, $P(L>X)$ and 95 percents confidence interval $p-$ and $p+$
}
\description{
loss probability distribution $P(L>x)$ by MC simulation
}

