% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SubSolver.R
\name{Projection_Main}
\alias{Projection_Main}
\title{projection method for soloving $min norm(Ax-b)$}
\usage{
Projection_Main(matA, vecb, lambda_init = 0, TOL = 1e-04, MAXITER = 1000)
}
\arguments{
\item{matA}{matrix A, matrix in (n,d+1) dimensional space}

\item{lambda_init}{= 0, set initial lagrange multiplier}

\item{TOL}{= 1e-4, set termination tolerance}

\item{recb}{vector b, vector in n dimension}

\item{MATITER}{= 1e+3, set iteration limitation}
}
\value{
optimal minimum
}
\description{
projection method for soloving $min norm(Ax-b)$
}

