% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SuppFuns.R
\name{psitheta}
\alias{psitheta}
\title{function value $phi(theta)$ for Two-steps Importance Sampling}
\usage{
psitheta(theta, x, ck, pkZ)
}
\arguments{
\item{theta}{exponential twisting parameter, postive real value}

\item{x}{threshold of portfolio, postive real value}

\item{ck}{expoure of obligors, vector in m dimenstion}

\item{pkZ}{default probability conditional on Z, postive real value}
}
\value{
function value $phi(theta)$
}
\description{
function value $phi(theta)$ for Two-steps Importance Sampling
}

