% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SubSolver.R
\name{sdp_solver}
\alias{sdp_solver}
\title{SDP solver for $norm(Ax-b)$}
\usage{
sdp_solver(matA, recb)
}
\arguments{
\item{matA}{matrix A, matrix in (n,d+1) dimensional space}

\item{recb}{vector b, vector in n dimension}
}
\value{
optimal matrix X and it's rank
}
\description{
SDP solver for $norm(Ax-b)$
}

