% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SubSolver.R
\name{sdprecovery_x}
\alias{sdprecovery_x}
\title{recovery vector from optimal matrix solution by SDP solver}
\usage{
sdprecovery_x(matX)
}
\arguments{
\item{matX}{matrix X, matrix in (d+2,d+2) dimensional space}
}
\value{
approximately recoveried x
}
\description{
recovery vector from optimal matrix solution by SDP solver
}

