% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SuppFuns.R
\name{twisting_probability_psitheta}
\alias{twisting_probability_psitheta}
\title{twisted probability for Two-steps Importance Sampling}
\usage{
twisting_probability_psitheta(theta, x, ck, pkZ)
}
\arguments{
\item{theta}{exponential twisting parameter, postive real value}

\item{x}{threshold of portfolio, postive real value}

\item{ck}{expoure of obligors, vector in m dimenstion}

\item{pkZ}{default probability conditional on Z, postive real value}
}
\value{
twisted probability by optimal $theta$
}
\description{
twisted probability for Two-steps Importance Sampling
}

