% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SuppFuns.R
\name{twisting_psitheta_equation}
\alias{twisting_psitheta_equation}
\title{function $phi'(theta) - x$ for Two-steps Importance Sampling}
\usage{
twisting_psitheta_equation(theta, x, ck, pkZ)
}
\arguments{
\item{theta}{exponential twisting parameter, postive real value}

\item{x}{threshold of portfolio, postive real value}

\item{ck}{expoure of obligors, vector in m dimenstion}

\item{pkZ}{default probability conditional on Z, postive real value}
}
\value{
function value $phi'(theta)-x$
}
\description{
function $phi'(theta) - x$ for Two-steps Importance Sampling
}

