% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SubSolver.R
\name{xy_lambda}
\alias{xy_lambda}
\title{Subfunction for Projection method}
\usage{
xy_lambda(matA, vecb, lambda)
}
\arguments{
\item{matA}{matrix A, matrix in (n,d+1) dimensional space}

\item{lambda}{lagrange multiplier}

\item{recb}{vector b, vector in n dimension}
}
\value{
list of x(lambda) and y(lambda)
}
\description{
Subfunction for Projection method
}

